Add support for lines with dual processes. (iOS Simulator)
